+++ src/django_ray/logging.py
@@ -146,3 +146,4 @@ def configure_default_logging(level: int = logging.INFO) -> None:
         handler.setFormatter(formatter)
         logger.addHandler(handler)
         logger.setLevel(level)
+        logger.propagate = False